# bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "AllungaWebAPI, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/api/ReportParams/id": {
      "get": {
        "tags": [
          "ReportsParamEP"
        ],
        "operationId": "GetReportsParam",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "SeriesID",
            "required": true,
            "type": "integer",
            "format": "int32",
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/ReportParam"
              }
            }
          }
        }
      }
    },
    "/api/ReportParams/{id}": {
      "put": {
        "tags": [
          "ReportsParamEP"
        ],
        "operationId": "UpdateReportsParam",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "SeriesID",
            "required": true,
            "type": "integer",
            "format": "int32",
            "collectionFormat": "multi"
          },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/ReportParam"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/TechReportEquipJobsPerDay": {
      "get": {
        "tags": [
          "ReportsPerDayEP"
        ],
        "operationId": "GetReportsPerDay",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/ChartDteTme"
              }
            }
          }
        }
      }
    }
  },
  "definitions": {
    "ChartDteTme": {
      "type": "object",
      "properties": {
        "ky": {
          "type": "string"
        },
        "date": {
          "format": "date-time",
          "type": "string"
        },
        "cnt": {
          "format": "int32",
          "type": "integer"
        },
        "title": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "ReportParam": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "reportid": {
          "format": "int32",
          "type": "integer"
        },
        "paramid": {
          "format": "int32",
          "type": "integer"
        },
        "deleted": {
          "type": "boolean"
        }
      },
      "additionalProperties": false
    }
  }
}